fix pepegas

+++ Models/ItemDrop.cs
@@ -1,3 +1,3 @@
 namespace Bot.Models;
 
-public record ItemDrop(string place, string item, string rarity, float chance);
+public record ItemDrop(string place, string item, string rarity, double chance);
